+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 2024-12-02 v1.16.9
 
 - Fix: Resolve the issue when the API is not using API keys.
+- New: Add `Enable .github\copilot-instructions.md` option.
 
 2024-12-01 v1.16.7